requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
sentence-transformers==2.3.1
numpy==1.24.3
torch==2.1.2
transformers==4.36.2
pinecone-client==3.0.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
redis==5.0.1
httpx==0.26.0
pydantic[email]==2.5.0